Add support to skip option parsing
When defining a command that takes arguments that look like options, it
would be nice to have a simple way to skip the option parsing logic and
just return all the arguments. This can currently be worked around with
the positional separator `--`, but this isn't a great experience when
you have to have that as the first argument every time.

This PR introduces the `skip_option_parsing!` method to the command DSL
which turns off the option parsing for that command.